CO2-forcing-1751-2014-demo/requirements.txt
numpy
pandas
xarray
netCDF4
matplotlib
pyyaml
